refactor: move aggregate_root from users to pkg
