+++ go/pkg/ddd/aggregate_root.go
@@ -1,4 +1,4 @@
-package domain
+package ddd
 
 import "github.com/google/uuid"
