+++ go/services/users/domain/user/entity/user.go
@@ -1,7 +1,7 @@
 package user
 
 import (
-	"GROWTHSPHERE/services/users/domain"
+	"GROWTHSPHERE/pkg/ddd"
 	valueobject "GROWTHSPHERE/services/users/domain/user/value_object"
 	"errors"
 
@@ -10,13 +10,13 @@ import (
 
 // User represents a user entity.
 type User struct {
-	domain.AggregateRoot
+	ddd.AggregateRoot
 	FullName valueobject.FullName
 }
 
 // UserEntity defines methods for user-specific behavior.
 type UserEntity interface {
-	domain.AggregateRootMethod
+	ddd.AggregateRootMethod
 }
 
 // NewUser is a factory function to create a new User entity.
@@ -32,7 +32,7 @@ func NewUser(id uuid.UUID, fullName valueobject.FullName) (*User, error) {
 
 	// Return the constructed User
 	return &User{
-		AggregateRoot: domain.AggregateRoot{Id: id},
+		AggregateRoot: ddd.AggregateRoot{Id: id},
 		FullName:      fullName,
 	}, nil
 }
